requirements.txt
torch
sentence-transformers
faiss-cpu
pypdf2
transformers
flask
